books_requirements.txt
ebooklib>=0.18
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
